texk/upmendex: Remove unused local variables
git-svn-id: svn://tug.org/texlive/trunk/Build/source@39650 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/upmendex/qsort.c
@@ -118,8 +118,8 @@ qst(char *base, char *max)
     register int ii;
     register char c;
     void *tmp;
-    int lo;
-    int hi;
+    size_t lo;
+    size_t hi;
 
     lo = max - base;	/* number of elements as chars */
     do {
